services/order/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
confluent-kafka==2.3.0
psycopg2-binary==2.9.9
prometheus-client==0.19.0
python-multipart==0.0.6